requirements.txt
streamlit
pytesseract
pandas
openpyxl
pdf2image
pillow